Adds missing docs, adds LocalUserDataStateService protocol, updates tests accordingly.
