+++ BitwardenShared/Core/Vault/Models/Data/UserKeyData.swift
@@ -1,6 +1,7 @@
 import BitwardenSdk
 import Foundation
 
+/// A type mirroring `BitwardenSdk.LocalUserDataKeyState`. Used for serialization.
 public struct UserKeyData: Codable, Equatable {
     let wrappedKey: EncString
